+++ framework-crates/objc2-thread-network/src/lib.rs
@@ -1,0 +1,22 @@
+//! # Bindings to the `ThreadNetwork` framework
+//!
+//! See [Apple's docs][apple-doc] and [the general docs on framework crates][framework-crates] for more information.
+//!
+//! [apple-doc]: https://developer.apple.com/documentation/threadnetwork/
+//! [framework-crates]: https://docs.rs/objc2/latest/objc2/topics/about_generated/index.html
+#![no_std]
+#![cfg_attr(docsrs, feature(doc_auto_cfg))]
+// Update in Cargo.toml as well.
+#![doc(html_root_url = "https://docs.rs/objc2-thread-network/0.3.0")]
+
+#[cfg(feature = "alloc")]
+extern crate alloc;
+
+#[cfg(feature = "std")]
+extern crate std;
+
+#[cfg(not(doctest))] // Some code samples are broken up
+mod generated;
+#[cfg(not(doctest))]
+#[allow(unused_imports, unreachable_pub)]
+pub use self::generated::*;